uint

+++ dedupe/core.py
@@ -57,16 +57,17 @@ def randomPairs(n_records: int, sample_size: int) -> IndicesIterator:
     else:
         try:
             random_pairs = numpy.array(random.sample(range(n), sample_size),
-                                       dtype=numpy.int)
+                                       dtype=numpy.uint)
         except OverflowError:
             return randomPairsWithReplacement(n_records, sample_size)
 
     b: int = 1 - 2 * n_records
 
     i = (-b - 2 * numpy.sqrt(2 * (n - random_pairs) + 0.25)) // 2
-    i = i.astype(int)
+    i = i.astype(numpy.uint)
 
     j = random_pairs + i * (b + i + 2) // 2 + 1
+    j = j.astype(numpy.uint)
 
     return zip(i, j)
